This commit finalizes some working changes to PPO code, namely adding learning rate and entropy weight decay
