+++ src/model_A2C_OD_PPO/.gitignore
@@ -1,0 +1,7 @@
+/logs
+/model_checkpoints
+*.py~
+*.py.un~
+*.py.swp
+*.un~
+/weights
